Split out tag creation into a helper class
