+++ src/Contrib/Jaeger/SpanConverter.php
@@ -8,10 +8,9 @@ use Jaeger\Thrift\Log;
 use Jaeger\Thrift\Span as JTSpan;
 use Jaeger\Thrift\SpanRef;
 use Jaeger\Thrift\SpanRefType;
-use Jaeger\Thrift\Tag;
-use Jaeger\Thrift\TagType;
 use OpenTelemetry\API\Trace\SpanKind;
 use OpenTelemetry\API\Trace\StatusCode;
+use OpenTelemetry\Contrib\Jaeger\TagFactory\TagFactory;
 use OpenTelemetry\SDK\AbstractClock;
 use OpenTelemetry\SDK\Trace\EventInterface;
 use OpenTelemetry\SDK\Trace\LinkInterface;
@@ -197,87 +196,10 @@ class SpanConverter implements SpanConverterInterface
     {
         $tags = [];
         foreach ($tagPairs as $key => $value) {
-            $tags[] = self::buildTag($key, $value);
+            $tags[] = TagFactory::createTag($key, $value);
         }
 
         return $tags;
-    }
-
-    private static function buildTag(string $key, $value): Tag
-    {
-        return self::createJaegerTagInstance(
-            $key,
-            self::convertValueToTypeJaegerTagsSupport($value)
-        );
-    }
-
-    private static function convertValueToTypeJaegerTagsSupport($value)
-    {
-        if (is_array($value)) {
-            return self::serializeArrayToString($value);
-        }
-
-        return $value;
-    }
-
-    public static function createJaegerTagInstance(string $key, $value) //TODO - split this out into a helper class to avoid the weird dependency HttpSender now has on this class
-    {
-        if (is_bool($value)) {
-            return new Tag([
-                'key' => $key,
-                'vType' => TagType::BOOL,
-                'vBool' => $value,
-            ]);
-        }
-
-        if (is_integer($value)) {
-            return new Tag([
-                'key' => $key,
-                'vType' => TagType::LONG,
-                'vLong' => $value,
-            ]);
-        }
-
-        if (is_numeric($value)) {
-            return new Tag([
-                'key' => $key,
-                'vType' => TagType::DOUBLE,
-                'vDouble' => $value,
-            ]);
-        }
-
-        return new Tag([
-            'key' => $key,
-            'vType' => TagType::STRING,
-            'vStr' => (string) $value,
-        ]);
-    }
-
-    private static function serializeArrayToString(array $arrayToSerialize): string
-    {
-        return self::recursivelySerializeArray($arrayToSerialize);
-    }
-
-    private static function recursivelySerializeArray($value): string
-    {
-        if (is_array($value)) {
-            return join(',', array_map(function ($val) {
-                return self::recursivelySerializeArray($val);
-            }, $value));
-        }
-
-        // Casting false to string makes an empty string
-        if (is_bool($value)) {
-            return $value ? 'true' : 'false';
-        }
-
-        // Floats will lose precision if their string representation
-        // is >=14 or >=17 digits, depending on PHP settings.
-        // Can also throw E_RECOVERABLE_ERROR if $value is an object
-        // without a __toString() method.
-        // This is possible because OpenTelemetry\Trace\Span does not verify
-        // setAttribute() $value input.
-        return (string) $value;
     }
 
     private static function convertOtelEventsToJaegerLogs(SpanDataInterface $span): array
